Refactored unit tests so only tests which test internals of the GetOpt class are in a test that inherits from GetOpt.
All other tests create a new instance of GetOpt before running.
